hw12/django/requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.10
django-celery-results==2.2.0
django-elasticsearch-dsl==7.2.1
djangorestframework==3.12.4
elasticsearch==7.16.0
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
idna==3.3
kombu==5.2.2
lxml==4.6.5
oauthlib==3.1.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
